Setup local env when running Docker for Mac.
Docker for Mac uses slightly different ownership permissions than regular Docker. It reports that the UID of a bind mounted file is that of the user who is checking the permission, and persists changes. So we need to change our assumptions about what the UID will be for bind mounted directories.